fix(be): 전적 데이터 수정

+++ user-service/src/main/java/com/a508/userservice/user/data/RecordRes.java
@@ -17,5 +17,5 @@ public class RecordRes {
 
 	private int rank;
 
-	private int odd;
+	private double odd;
 }
